feat: :art: texture in stack, fix minimap dir, norminette, etc ..
+ world deinit added, backroom texture added, g_variable all deleted.
