+++ backroom/ceiling0.xpm
@@ -1,0 +1,85 @@
+/* XPM */
+static char *_650504336711[] = {
+/* columns rows colors chars-per-pixel */
+"64 64 15 1 ",
+"  c #E9E6C8",
+". c #EAE7C9",
+"X c #EAE7CC",
+"o c #EAE7CD",
+"O c #EAE8CD",
+"+ c #EBE8CD",
+"@ c #EBE8CE",
+"# c #ECE9CF",
+"$ c #ECEAD1",
+"% c #EDEAD2",
+"& c #F1EFDB",
+"* c #F2EFDE",
+"= c #F2F0DC",
+"- c #F3F1DD",
+"; c #F3F1E0",
+/* pixels */
+"****************************************************************",
+"****************************************************************",
+"*******************=********************************************",
+"*******************=********************************************",
+"====OOOOo#o#############$o#o#o#o#o#ooOOOOOOOOOOOOOOOOOOOOOOO----",
+"====OOOO#o#o############o#o#o#o#o#o#OOOOOOOOOOOOOOOOOOOOOOOO----",
+"====OOOOo#o##############o#o#o#o#o#oOOOOOOOOOOOOOOOOOOOOOOOO----",
+"====OOOO#o#O############o#o#o#o#o#o#OOOOOOOOOOOOOOOOOOOOOOOO----",
+"****$$$$$$$$$$$$$$$$$$$$$$$$$$$$#o#o#o#ooOOOOOOOOOOOOOoo##@@****",
+"****$$$$$$$$$$$$$$$$$$$$$$$$$$$$o#o#o#o#OOOOOOOOOOOOOOo@@@@@****",
+"****$$$$$$$$$$$$$$$#$$$$$$$$$$$$#o#o#o#oOOOOoO@@OOOOOO@@@@@@****",
+"**==$$##$$$$$$$$$$$#$$$$$$$$$$##o#o#o#o#OOOOo@@$OOOOOO@$@@@@****",
+"====++++++++o#oO++++XXXXoooo@@@@$###$$$$$$%%%%%%$$$$%%%%$$$$****",
+"====++++++++#oX#++++XXXXoooo@@@@####$$$$$$%%%%%%$$%%%%%%$$$$****",
+"====++++++++++#X#X#XXXXXoooo@@@#####$$$$$$%%%%%%%%%%$$$$$$$$****",
+"====++++++++++++X#X#XXXXoooo@@@#####$$$$$$%%%%%%%%%%$$$$$$$#****",
+"====oooooooooooooooooooOOOOOXXXXoooo$$$$%%%%%%$$%%%%$$$$$$$$;***",
+"====ooooooooooooooooooOOOoOOXXXXoooo$$$$%%%%%%$$%%%%$$$$$$$$****",
+"====oooooo@@ooooooooOOOO@X@@XXXXoo@@$$$$%%%%%%$$%%%%$$$$$$$$****",
+"====ooooooo$ooooooooOOOOX@@@XXXXoo#$$$$$%%%%%%$$%%%%$$$$$$$$****",
+"----$$$$%%%%%%%%%%%%$$$$@@@@@@@@$$$$$$$$$$$$$$$$$$$$$$$$####*&-&",
+"----$$$$%%%%%%%%%%%%$$$$@@@@@@@@$$$$$$$$$$$$$$$$$$$$$$$$####*-&-",
+"----$$$$%%%%%%%%%%$$$$$$@@@@@@@@$$$$$$$$$$$$$$$#$$$$$$$$####-&-&",
+"----$$$$%%%%%%%%%%$$$$$$@@@@@@@@$$$$$$$$$$$$$$$#$$$$$$$$####&-&-",
+"****%%%%%%%%%%%%%%%%%%%%$$$$@@@@@@@@oooo++++++++++#Xoooo##OO=***",
+"****%%%%%%%%%%%%%%%%%%%%$$$$@@@@@@@@oooo++++++++++X#ooooXXOO****",
+"****%%$$%%%%%%$$%%%%%%%$$$$$@@@@@@@@oooo++X#++++++#Xoooo#OOO****",
+"****%%$$%%%%%$$$%%%%%%%$$$$$@@@@@@@@oooo+X##++++++oooooXXXOO****",
+"****oooo$@OO#o#o$$######@@@@@@@@########$$$$########o#o#OOOO----",
+"****ooooooOOo#o##$######@@@@@@@@########$$$$#########o#oOOOO----",
+"****ooooOOOO#o#o########@@@@@@@@########$$$#########o#o#OOOO----",
+"****ooooOOOOo#o#########@@@@@@@@########$$$##########o##OOOO----",
+"****O#X##################o#o#o#o@@@@########$$$$$$$$$$$$####----",
+"****#X#X################o#o#o#o#@@@@########$$$$$$$$$$$$####----",
+"****O#X##################o#o#o#o@@@@########$$$$$$$$$$$#####----",
+"****#o#+################o#o#o#o#@@@@########$$$$$$$$$$$#####----",
+"&&&&    ........XXXX++++oooooooooooooooooooo++++XXXX#X++OOOO----",
+"&&&&    ........XXXX++++oooooooooooooooooooo++++XXXXX#++OOOO----",
+"&&&&    .......XXXXX++#XooooooooooooooooooOX++#XXXXX++++OOOO----",
+"&&&&    .......XXXXX++X#ooooooooooooooooooOX++X#XXXX++++OOOO----",
+"&===XOOOOOOOOOOOXXOOOOOOOOOO#o#oooooXXXX#X#XOOOOoooooooo#O#o----",
+"====OOOOOOOOOOOOOOOOOOOOOOOOo#o#ooooXXXXX#++OOOOooooooooO#o#----",
+"====OOOOOOOOOOOOOOOOOOOOOOOO#o#oooooXXXX++++OOXXooooooOO#o#o----",
+"====OOOOOOOOOOOOOOOOOOOOOOOOo#o#ooooXXXo++++OOXXooooooXXo#o#----",
+"====########$$$$$$$$####OOOOOOOOooooooOO##@@@@@@OOOO........&&&&",
+"====########$$$$$$$$####OOOOOOOOooooooOO@@@@@@@@OOOO........&&&&",
+"====########$$$$$$$$####OOOOOOOOooooOOOO@@@@@@@@OOOO......X &&&&",
+"====########$$$$$$$$####OOOOOOOOooooOOOO@@@@@@@@OOOO...... X&&&&",
+"----OOOOoXXXX ..++++++++++++ooooOOOOOOX@OOOOooooXXXX........=&&&",
+"----OOOOXXXX X..++++++++++++ooooOOOOOOOXOOOOooooXXXX........=&&&",
+"----OOOOXXXX....#X#X++++++++OoX#OOOOOO@@OOOOooooXXXX.. +....&&&&",
+"----OOOOXXXX....X#X#++++++++XX#XOOOOOO@$OOOOooXXXXXX +  ....&&&&",
+"=***#o#oo#O##$$$$$$$$$%%$$%%%%%%$$$$%%%%####++++#X++++++XXXX====",
+"****o#o##o#O#$$$$$$$$$##%%%%%%%%$$%%%%%%####++++X#++++++XXXX====",
+"****#o#oo#O#$$$$$$$$$$$$%%%%%%%%%%%%$%$$####++++++++++++XXXX====",
+"****o#o##o##$$$$$$$$$$##%%%%%%%%%%%%$$$$####++++++++++++XXXX====",
+"****#o#oo#o#####@@@@  ..#XX#oooo$@OOOOOOOOOO$o++XXooooooXXOO----",
+"****o#o##o#o####@@@@ X..X@@Xoooo@@OOOOOOOOOOo#++XXXoooooXXOO----",
+"****#o#oo#o#####@@@#....++@@ooooXOOOOOOOOOOO#@X+XXooooooXXOO----",
+"****o#o##o#o####@@@#....++@$oooo@XOOOOOOOOOoo@#XXXooooooXXOO----",
+"====----********************;;;;;;;;-;;;;;;;****************=-=;",
+"====----********************;;;;;;;;-;;;;;;;****************==*=",
+"====----********************;;;;;;;;;;;;;;;-********************",
+"====---=********************;;;;;;;;;;;;;;;-********************"
+};
