+++ backroom/floor0.xpm
@@ -1,0 +1,99 @@
+/* XPM */
+static char *_650504340384[] = {
+/* columns rows colors chars-per-pixel */
+"64 64 29 1 ",
+"  c #E5E3C1",
+". c #E6E3C1",
+"X c #E6E4C1",
+"o c #E6E4C2",
+"O c #E7E4C2",
+"+ c #E7E5C3",
+"@ c #E8E5C3",
+"# c #E8E6C3",
+"$ c #E8E6C4",
+"% c #E9E7C4",
+"& c #E9E7C5",
+"* c #EBE8C5",
+"= c #EBE9C5",
+"- c #EAE8C6",
+"; c #EBE9C7",
+": c #ECE9C7",
+"> c #EDEBC7",
+", c #EDEAC8",
+"< c #EDEBC8",
+"1 c #EEEBC8",
+"2 c #EEEBC9",
+"3 c #EEECC8",
+"4 c #EEECCA",
+"5 c #EEEDCA",
+"6 c #F0EECB",
+"7 c #F1EECB",
+"8 c #F1EFCB",
+"9 c #F1EFCC",
+"0 c #F2EFCC",
+/* pixels */
+"%%%% ooo##%%3111oo-o-oo--%-%@%@--o--4494%-ooooO ::%:11111434>1%1",
+"%%%%oooo#-%%3>41o-oo-o-o--%-%@%@--o-4419-%;- o  ::%:11113343>-1;",
+"%%%%oo  %%%%1414ooooo-o--%-%@%@O@o--4444%%%-  o >%::11>1343441%1",
+"%%%%ooo %%%%4141%o@o-o-o-%%1%@%O---@1919%%%# o o:::%11114343%%4-",
+"%%%-    :11%1411oooo    o    o  4@4@1414         %oooo  o%%o@%@%",
+"%%%%    :-114111oooo        o @ @1@14141        o %ooo  oooo@@% ",
+"%-%%    1-4-1414oooo  o      o o:@4@1414        %%oo    %o %%o%%",
+"-%-%    ::-14141oooo  oo    o @ @6@14141      ooooo%    %o% %%%o",
+"ooo%o%%o%%o---11-4-4-%%%%%#-oooo@@--6666:6::--44%11%#o# %%-%-:-:",
+"o-oooooo%%-%411:----%%%%%%%#oooo---@66666::14>>-1%%1##o#%%-o----",
+"oo%o%oo-%-%-4;11-4--%%%%#-##oooo---@66661;114-;;o4%%####-%-->%-1",
+"o;oo%ooo-%-->11;--4-%%%%####ooooo@--6666411:;-4;1@1%#o##%-o-%>-%",
+"ooooo-o-    ooooo----@@@1:>:;--3--2@:661;;1-11144:::%%%%X#X###o#",
+"oooo-o-o    oooo-oooo@--:>:>;;---2@2:4:61;%111;1:11:%%%%#X#X# #o",
+"ooooo-o%    oooo-%%o@-@@11>#;;>--@2@6:6:;-1-;;11>>>>%%##X#X#X#o#",
+"oooo-o-o  oooooooo--@o@@@@11;;---2@2:6:6;1;1111;>>>>%%###X#X#X#X",
+"o-oo;;;;;%;%;%%:%%#O:::>499843>3----4411o@@-1-1- O%#     %o- ooo",
+"oooo;;;%%;%;1%%:%oO#::>:99483-33----4111--o@@4-4%%##    % o%oooo",
+"o;oo%;%;%%-%%%%%% %O:>:>949433------3311@o@@411-####    oooooooo",
+"ooo-;%;%%1%>%-:%o%O#>:>:9495>343--%-3311@-@@4@-4####    o%-ooooo",
+"oooo%;%;111111>6%%%%:%-%1;1;;;;;-- %4%%1o-oo@%oo:%;---%%%%%%    ",
+"oooo;%;%11114141%%1%%1%1;;;1;;;;-#%%%-%%@ooo%@%%%1%-ooo%%%%%    ",
+"oooo-;%;11>>33>41%1%1%:%;1;1;;;;-% %-%--oooo@%@%-%1%%o%%%%%%    ",
+"oooo;o;%11133361%%%%%%1:;1;;;;;;o %%%-%ooooo%@%%%-%-o%%o%%%%    ",
+"o   o%oo;;;4>>:1%%%%6@46:6:6;%;%oooooo%o #  %%@@%:--3>->o%ooo   ",
+"    o-o%1;3->>1:%%%%%@@%6:6:%-%; ooo;oo%O  %Oo@@:%>---44oooooo  ",
+"    -ooo-3-3->-3%%%%@26@22:6-%%-oooo-o%o OOO%O%@*>-;-3---%oo    ",
+"    oooo3-3->>3-%%%%6%@2222:%1-:ooooo-o-O OO %O@%#>-3->3ooo-    ",
+"        -o-o-3-3-%@@1%2@26:6:@-%%%%%>>>%>>>>6@O661114343-;;;o%-%",
+"        o-o-3-3-%14@@4@:6:6:%:%-%%%%>-:1>1>>O:6%>>113434-;;;%%%-",
+"        -o-o-3-3%4@@:@:@:6:6-%-%%%%%::::::::6:#611114411;%-;-o--",
+"        o-o-3-3-%%@1@6@66:6:%:%:%%%-:::::::26O:611114111-1----%-",
+" %## O o #X#-->-%;-;%%o;444-519411>4--%;3--%88998338--o--3146666",
+"% # O   O #o>>%#;-;;%%;o;;44591411>3>%;o--3-88994843-o-->3416696",
+" % # #% OOo#--->%;;o%%%%44449144%;44%%%;----89998484---->4146069",
+"% ##    OO#o->%-;%;;%%%%4;411941114-;;%%->--99994384-o-o41416696",
+"%%% %o%%  #             o   %;%;--;;%%%;o-oo84443434-%o-%16%66*1",
+"%%%% %oo# #             o   ;%;%>1%%%;;oo---34844343-oo-1%@@:>11",
+"-o-o# ## #                  %;;;--;;;oo;o%--48443434oooo@1%6:111",
+"o-o-# ##   #                ;%;;->%;;;o;-%o-44434343oo-o4@%:>>:1",
+"%-1@@o@o#X#XXX          %ooo1;1%3-3-%>%-#%%%43-34444-----1-4@@-@",
+"4-@4o@o$X#X#XX          o#%o;;;;-3-3#>%--#%%;1434444----4-4--@-@",
+"%%4@$o$o#X#XX           #o#o;1;13-3-#%>%%%%%11334444-----4-4@-@@",
+"%1@1o$o$X#X#X           o#oo;;;;-3-33##%%%%%>>334444----::::@@@-",
+":114$ooo#o#X#####>#-oooo--%-4694->-%%o#X3-%3>----%%%-ooo-@@o@   ",
+"11--ooooo#X####o-o-%oooo3=>34444%-3%o#X#3%-3%3>>-%%%ooo-oo--@   ",
+"1-11oooo#o#o##o-o;--%o%o----831433--#X#X--%%3>>3%%%%-o-o@-@@    ",
+"4411ooooo#oo###o--#-ooo%-3-34466-%%3X#X#%33-=3--%%%%oooo%o@@    ",
+":-1o   o    %%;%-1-1-%;%191500001%%1oooo%%%%->-3-%-%oo-%o OO    ",
+"-%-,oo o    1%;%::-%;-%;94940000%1%:oooo%%%%>>3-%-%-%oooooOO    ",
+"----ooo     %;%%:%#6%%;%19190086%4:%oooo%%%%>>>3%14% %o%o@ O    ",
+"----oooo    %%%;:%1%-;%;945100961@%1oooo%%%%1>33%%%%% %o@o@O    ",
+"----o-oo     ooo#O%@3341598866364141% %o #O -%--%-%:oooo        ",
+"----  oo     o  O#O@4444998864641-4% % %   O-%%%-%--oooo        ",
+"----   o        #O#-3451966836362::6% % # # -%o1%-%-oooo        ",
+"----& oo        O###434-9988636326@: % %  # --;---:%oooo        ",
+"##%%oooo            ---,-o%-o%%o..@.@@o@-%%->>3-:::%%%%1@ o%oooo",
+"#%%%oooo            --,---oooooo...@@@$o%->#334;%%::::%%@%o%oooo",
+"%#%%oooo           o--1,oo-o%oo%@.@.o$o$#->-33>1::%:%%%%o@oooooo",
+"#%%ooooo           -1o--o-o-%o@o.@.@$o$o----33>1:%-:%:-%@o%ooooo",
+"6161-3-31%%%    oo-o-oo--ooo%%%%:%%:42-284444111%O%#% % o-o-%o@o",
+"66163-3-%%--    o-oo--o-ooooO% %:::%224:43831%11@%##o% %-o-oooo-",
+"9191-4>>%-%%    oooooooooooo%#%#%:---1:148441111####ooo o-o--ooo",
+"19161->1%%>#    o-o-o-o-ooooO%%O%:-:11:149444->:####o%o -o--ooo-"
+};
