+++ backroom/floor1.xpm
@@ -1,0 +1,111 @@
+/* XPM */
+static char *_650504343913[] = {
+/* columns rows colors chars-per-pixel */
+"64 64 41 1 ",
+"  c #D9D7B4",
+". c #DAD8B4",
+"X c #DBD8B5",
+"o c #DBD9B5",
+"O c #DBD9B6",
+"+ c #DCD9B6",
+"@ c #DCDAB6",
+"# c #DCDAB7",
+"$ c #DDDBB7",
+"% c #DDDBB8",
+"& c #DEDCBA",
+"* c #DFDDBB",
+"= c #E0DEBA",
+"- c #E0DEBB",
+"; c #E1DEBB",
+": c #E1DEBC",
+"> c #E1DFBD",
+", c #E2DFBC",
+"< c #E2E0BC",
+"1 c #E2E0BE",
+"2 c #E3E1BE",
+"3 c #E4E2BF",
+"4 c #E4E2C0",
+"5 c #E5E3C1",
+"6 c #E6E4C2",
+"7 c #E8E5C3",
+"8 c #E8E6C3",
+"9 c #E8E6C4",
+"0 c #E9E7C4",
+"q c #E9E7C5",
+"w c #EAE7C4",
+"e c #EAE8C4",
+"r c #EBE8C5",
+"t c #EBE9C5",
+"y c #EBE8C6",
+"u c #ECE9C7",
+"i c #EDEAC7",
+"p c #EDEBC8",
+"a c #EEEBC9",
+"s c #EEECC9",
+"d c #EFEDCA",
+/* pixels */
+"00006495tt00aaaa855588581881333333338888>>5>>>,=6608iiiidddduuuu",
+"00006446t808aaaa558586888338333333337383,6>>**=,0088iiiiudduuuuu",
+"00006666888taaaa5858885588883355333338884*1,,,,=8888iiuudddduuuu",
+"000066658te8aaaa585658881381333,33338883,1,,*1*,8888uuuaduuduuuu",
+"00qq64440uu0ta00,111%%%%@@####X%=@=@=@=@  %%@%@ ==%%,,3166668888",
+"00qq44440000t00a11=3%%%%@@####X%X=X,==%=      @@=@=%,51366668888",
+"000q4444u0a0yyyy1,,1%%##@%@% @@X,X, %,; %   %@@ %=%%141166668888",
+"000q4444uu00yyyy1*41%%##%@%o%oX@ =X=%%%=  @ % @@*@%=4,1366688888",
+"95654646=,1<<*<<%%%%%###XX@X####%@X,==@=@@=@=@=@.<&*&&o<888888ue",
+"999t6464,,*1*<*<%%%%## #%%X%####=X@X=@==@=@=@=@==o&o<&<<86580tee",
+"59664646,*,*<&<*%X%%  ##XX% ####%*;X=====@=@=@=@o=o<&&&@686eeeee",
+"59666466*4*1<<<<%%%%### X% o####++*%=@==@=@=@=@==o<.@<<<8683eeee",
+"6645*1**.&o...oo %% X%@X==X=####+*####@@@@###*@@o=@=@@&@=*=*6666",
+"6655*1*1oo....oo % % XX@@%= ####++#X##*%o%##@@=@=@=@@&@@,*,,6683",
+"5555****ooo&&o&.% % %@.X.=X=####** *###o%###@@@@@=@=&@o***==3838",
+"55551***#o.@...&X %X% ..=.=.#### X* *%*%####=@=@=@=@o&&o,===8383",
+"4*1<oo*o=%=@@@o@@oo.==..@=@=@@X%##o%@@%@@@@X@@@%.@@@%@  %%%%1411",
+",4<1#*o*@=@=@%=ooo.o.o.==@=@%@%%##%@=@*@.@ @@@@X.@@@ @%@X%%%4,13",
+"4*41*o*o=@=@*&@&.oo....=%=@=@%%%##@@@@@=@.@ @@%%@@@@ %%@%%%%3131",
+"*411o*o*@=@=@@&@.o.o==..%%=%% % ##@@=@%@.@  @@X;@@@@@ @@%%%%1313",
+"*****o*o<,,,<@@@XX=X..=.@@@@@@%Xo %X@@@@.... @  .@ X%%@@####===,",
+"****o*o*=<,,@@<<X XX=.=.@@@@@X%% %% @@@@.... X XX@%@%X;@####==,=",
+"*****o*o<<<<@<@@XX=XX...@@@@@@X%#%@ @@@@....  X.@XX %@;@####=***",
+"****o*o*<<<<@@<@XXXXX;..@@@@@@ ;  o%@@@@....@  .X@@%@%@@###o,*,=",
+"*o1o####****@<@@XXXX@@@@=X=.@ XX## #o @ ...o@.X @@@@@&@&### ****",
+"o*o*####*#**@<@<XX++%X@@X=X=XX% # ## @@@.@...@.X@@@@&@&@### **1*",
+"-o>*####*&**@@<@#X++X@X=X==X@X.X##  @@@ ....@.@.@@@@@&@@o###***%",
+"***o####*1o*<@@@XXX+@@ XXX=XX@XX ###.@ @...@.@.@@@@@&@@.#  #*%*,",
+"*o*ooo# @@##*o@<* @@%oX@=X.=X..=***,=,*=<<@@<.<.<&<&&<.<&.&.*111",
+"****o   @@##**@=##o%#%=@X==.X=.=*,o===*,@<<<.<.<&<@<<.<&.&.&1111",
+"*o**  ######&&@@###o* = .=.=....***@,=*=@<<&<.<@<@<<&<<&&.&.111*",
+"****o   ####o-=&#### =o==.=..==.&******,<<@<.<@@<<<@<@&<.&.&<<11",
+"1111#  # o o@@%o######%X@*@=*,,,**<@@&@@@@@@<<&<&<1<o<o**&&<3ttt",
+"1111####o o @@oo#######%=%=@***%&<<@@@&@&@=@&<<.<<<1*&*&<&&*tttt",
+"1111 ##  o  @@o####### **o%*,*,;<&<<.&.&@@&.@<.<1<1<o*.<&&<*rrrr",
+"1111 ###o oo@@o.######%%*====*=,&<@<&.&.&@o&<&<&<1<1<.<o@<.<rrrr",
+"6465%###   #        %    @  @@@@.&&o..@.& ..<&<&<*<<.@.&@<3@rrrr",
+"4936##*# %            %    @@@@@&ooo@@&o. & .<.<<*@<@@&.3<,<rrtt",
+"8383#**%# #   ..     %%    @X%X@.@o&.&...&.&<&<.&<&<@&.&,1<<tttt",
+"3838#%#*  #   ..        @X  @X@@@@#ooo.o.&.@.<&<<&<@&.&.@1<,tttt",
+"8tr8==@=oooooo..   oX % o%  ## @&o@&o*o##o&o<.<.**,@#%#%66388888",
+"88r3=@=@ooooo ..   % * % %##*#=..&@@o&o###&o.<.<***=#*#*66838888",
+"8088%=%=oo#o ...    ##%  @@@@* =@=@&oo&o@.oo<.<@*****%##38387788",
+"08y8%%=@ooo#o... %%   ##@oo@@%=o@@@@&oo.@&.&.<@<&&,*%o#*83837778",
+"8880@XX%########%%%%****====1=1<<<<@@@&.&.@@@=@@@@@@*###5553,555",
+"8880X;%%########%%%%,%,%====33<1@<<@@@o.@&=@@@=@@@%@#X*X5,555555",
+"8805%%X%########%%##****=====1<<<<<<.@@.@@@@@@@@##@%#*%%,5355555",
+"8800 % %########%%##,#,%=*,==3<<<@@<@..&@&@&=@=@##%o%%#%55555555",
+"5555 #####%%,11<%=@=%%%%=%,,<=<=*@%% oo..&.&o&&=####--@-188,7555",
+"5555## #%#%%,1<<=@=@=%=%=,%,=<=<@%@o ...&.&@<=@@####-o--810=5555",
+"5555# ##%%%%<<<<@=@=%=%=**,,<=1=@&#*o .@.&@@o&=&####****13885555",
+"5555% o##%%%<<<<=X=@=@%%,,%,=<==&@### ..&.&@@=@&####****88185555",
+"5791@%######=*%*@=@=@-o>*%=@=@=@@&@&& ...ooo&@==*,<*<11444555555",
+"5501%o######*%*%*o=@>*>*#*=*%=@=&@&@. & ...&=@&@****614144455555",
+"7,00########*%%*o<@-*-#>*o=%*@=@@&@&.& &&.@.=&&@=*<<<5,4445,5555",
+",00,########***%-@-@-@**,*==@=@=&@&@ &...@.&@&*&,***5,4144355555",
+"7788****# # #% @oo  **##%%%%%*%,=1,,,,4,6333tt880707000077606666",
+"7786**** ## o oo oo *##*%%%%,***<*,=,44*33e386889070000075606666",
+"7700****####ooooo o.o###%%%%**%,*<*=*4,1363688889y07000075566666",
+"7766****#  #oooo o .#**o%%o%*#*,1*1*14*4363688887700000050066666",
+"aaaseette331,=,**11131114116858500tqaaaudddduaua0707669999996660",
+"aasaeete<1ee**=*11*11311155,88050t6qaauaddddaaau7070609599996066",
+"ssaseete1e61<*,*1*<1314144,585880099psauddssuaya887766069900q606",
+"ssssettu3111**<*1*1<134*4,5,8588y0ttpusudddpsuaa88880666990q6606"
+};
