+++ backroom/floor2.xpm
@@ -1,0 +1,100 @@
+/* XPM */
+static char *_650504348600[] = {
+/* columns rows colors chars-per-pixel */
+"64 64 30 1 ",
+"  c #6D5E24",
+". c #6F5F2B",
+"X c #715F26",
+"o c #715E2B",
+"O c #6F6026",
+"+ c #6F6129",
+"@ c #746226",
+"# c #786627",
+"$ c #75652C",
+"% c #79662C",
+"& c #77682D",
+"* c #7A6A2E",
+"= c #766631",
+"- c #796632",
+"; c #776831",
+": c #7B6C33",
+"> c #7E6E39",
+", c #7F7035",
+"< c #7F7139",
+"1 c #806E2F",
+"2 c #816E34",
+"3 c #816E39",
+"4 c #827136",
+"5 c #83733B",
+"6 c #88763C",
+"7 c #87783E",
+"8 c #8A793E",
+"9 c #857641",
+"0 c #877841",
+"q c #897A41",
+/* pixels */
+":<;<;;;<;+@-X@::,:,:::%:@%%%@@@*#1*4844*2422*%:22246q622:24:56&:",
+":<:::;;;;&%:.@-3:,;<5:3%:%@1**@#2%**6824*252:2%%*232262625::5q,&",
+";$5<<;;;:<::;:53,,,,:%%:*:%-2**2****6642%2*2%222%:52525652:5,<,,",
+";+:<;+;;:;;:*-53,&,,%:%;2%:22*22**%:26**244:2222662225::502:,<,<",
+">=;<;;<+::&*:&;*,,,,&,&,%:66:2:4:1*%**44864:22:66q2256$&,,<<<:<:",
+">><;;=;;*::$::::,7,&&&&1%2242::2%::4:4226245::45662565*:,;<<<0<;",
+"<<>;<<;<*:$$446;,&7,,,&&$4**2%:**465*4%$552:*:*:66::2:45;;<<;;>>",
+"<<>;;<<<*:&@57461,,7,,&&**%*%6%*,44*2:@%:6542*::2224::::;<<;<;<>",
+"<<<<,<,,*464,&,,&#&&,&,&*&,*,,&&,*:$&,:4,*45<,**<<;,<<<0,;<;0;<;",
+"<<<;,,,,,547*#**#&&&&&,#*&**,,*#,4,*$,,,*$67<<**77,;<<<<;<<;;;<<",
+"<<<<;;&&45,*&&,,&&&,&&&,&,7,*,&,44*<44$,,:::6<::&&7;,<<;<;<<<<;<",
+",<,<;<,&,,,,*#7,1,,&&1,,&,,&#,,&,*:*66**4,<457::;,<<<;<<<;;0<<<<",
+"<<<;,,,,7&,&*,,,,,,,&1,,*#,,&,,,,4$&$*<656::65&<<<,7<;0<;<;<<;<<",
+"<;;;,,&,,,1&#&&#,&,1,&*,,,,*&,,,$4#&**4,57*5:5:<<7<<<;<<;<<<;;<;",
+"<<,,,,&&11#1&#&*,11,,,,,,&,*,,<7,,,,&,5,<*q55:55<;<<<9<0<<;;<<<<",
+"<,,;,&&&,&#O&#1#,&111,,,*,@&&,,<4,,4<$,4;4<4<<5:;<;<<<<<;<;;<;0<",
+";;;<<<<,+;;;+&&&&O,,,,1,>>=;<<0;6,@:,6:<5q5:5555<;0<99<<>==;>>;>",
+"<<<;<;;,.;;;&&&&&&,&,,,7>9>.<;<9,4$&54450q7<5;95;;;;<9;<<>==<>>;",
+"<;;;+&;,>;>><&&&&,,,,&,,=>.><;<<7;&:,:55577q<:>:<<<<===<9<<<><;;",
+"+<<;;,<,==9>;;&+,,,&,,&,>==>>9<;<::<*,>:5:07;;>5<<<<>.=;99<9;<<;",
+"<;<<<7;;=>>>>>9><;>>7,&,>>=>=9==:;::53>55<559<>>99=><>==;:>:<;;<",
+";<;;<<<<>=9=>>9>><<<,,;,>>===>==$:=:59q50>9<99;>99>9>>==<;>;<<>;",
+"<<<;;;;<>>>=99>>;;<;<0<<>>>===.=>9>;>>;355>;:>99>>>>=<<>><>;<<>:",
+"<<<<;;;;>===9>>;>>;<;<0<>>=>==.=>>=;>3:;95>:>-9>>><<999>;>:,<<<:",
+"<0<<<<;;<7<<9><>>>>>=>>>9>====>=>=>>9>=9=>>>=>>>>>q<<;;<;::::;&O",
+"0<<<<<;;<<<<><>>>=9>>>>>>9==.>.>>>>=>>>->>>>>=>=>>9<:<:,+:;:,:;+",
+";;++;<;&<<<<;><;=>>==>>>>>>>>>.=>9>>=>====>=9>>9>>::9<<,:<,:<5;,",
+"&;;;<<&&;<<0<<<<>=>>==>>99>=9>.=>>->>>=>>>=>9>>>=>;;<0,<:<:,:<;<",
+";&&,&;;&<<;;00<;>==>>>>>>>>9>.=>>>>>==>>==>=>>9>>>>=,$<:&&,&,:&$",
+"&&&&&;;&<<;<00;<>====>=9>9>9=>>9=>>>=>====>=9>99=>>>&;<<$&&$&;&+",
+"&&,,&&&;;;;<<<;<9>>>=>>=>>===>>===>o==>>9>:9>><<9>..:;$,,:,*$,:*",
+"&&,,&&;;;<++;;<<>9=>>>9>>>=>>.==>=>===>>>>>>>>9:><.=<,;<::,*,,5&",
+"*,@,$@,5;<:+;;;;;>;:;:><5<>:+:$;=>;><:999559<><:5<;;<577,,&&,,7,",
+",&&*&&54:;;+=+;;<<:>;:55<<;:+=;==>:;><00>55q><<<<5::55776,&#,,74",
+"*#$*,,,4&O&:;&&$>:$:;<<<<<:><;$:>>><05>5:77:77;:;<&$;,*$,,,,,&&&",
+"&&&&,,,,&O;&+=$:;>;::>05>:<5;:;&<>;<<<:<55<759;:&:+$,,,,,,,,,,,&",
+"&,,#,**$O$+&,,;;:&:;;&5&5<:;:;$:&$:;:<55,,<7&:5<,5<&55*,,,&<@,&*",
+"&&,,,,,&$&:&,:<:$:::::&$;<<:&:&;$$:,::4<<57<&::<<,,&<55<<,,<,&,,",
+"*&,7&*,$$$:*&<;:<:;:<::$&:@++$:<&;<,:<:<<555<5<<<5:5<<,5,,&&*,#,",
+"&#,**OOO$:;&::,,<0&::<$;:$OO:&:&;&:<5,6<,:<&,5,&:5,<:4,:,,&&&,&,",
+"**&;&&OO$:+&$+<7q75457:<&;:;;&<<<:<5<5<,55,,,q&$<,5<<,;,,;,:$;;&",
+"&&+&&&O ;O&$;$&<57,<q5:$&&&&:&:,754:,,5q,,,<5q;:,567,<:&,&,,;;,;",
+",,,&&,,&+$&$&$:$<5:<,&&:;$&,&::&$,:<;,$:*<:,,&,,&;,&,<77;,,:<,<,",
+"&&&&O&&;&&&$$$*:*:&,::$$&;&&:+:&,:*<*,$&&&::,,$,::&<4,5,&&;,<&<,",
+":&,;:;;&&&&*@*,,*$,$*,&O&&*,6,$%654,74&,<*,*6,$&*&,&*,$$&,,,&&$,",
+":<&:;$,,*&&:&,,*&@&**&$$$@&&$*$,,,5,,,,,,&,,,,,*:,*,&&@$4:&,&&&<",
+";+;;$&&;,&$,&O$$,,*,$&*&$O$&,,<<$*,**,&*<:**&,5,,&&,$@*,&,,<,<,,",
+"+&,;&&O&&***OO$&,4,,$&*$O$@*4**4**,,,,*&*4,$$**,*,&,$&,,,&,5,<,<",
+"<<;:=>::;:%*6<,*o11o122o122:%26$&*,*,4*4o-112232*&,:::4:,*4::::&",
+"<0$=:=5:*5;::4,*1o1ooo2o%o:2X2:*@$,4<*:4%-222231:**,2%:*&**;;<$$",
+";=;=>$--::55-5&*23232222%:#%111X&*1#1*###%123211#*44:422::;*3:7:",
+"><;$$>::;-5;%:$$31222o22%**%X1X1,**:@#*#%%222o11*44,362o:4$%3;5:",
+";<<;;;>=9-----3322o2o2o1%%%-21X1##11111#*1##1 11*#166413:%32$:;:",
+"::<>>;=>---o>-3-22223211%#1122X1#*#41#111### 111#:64461%%-3:;::>",
+"=;:;>53=->=-3-33o2123332%2#%11oX1111#111@##11 1166#*1122%X70--3>",
+"$;;:5>==>=3--33-332o3222*%%%1%o#11111#11##1*X 1161#*2262-%352>::",
+"@.;::3$----9o-3-323233321X221XX1##1#411*21211122112o1232666522-2",
+"$@$;33>o>-3>3333331o2322X121111X####*1111212112oX131oo2262333-::",
+"7:24$-$----o----2333oo2211ooXoX1#1**41111311o111121XX12211-32566",
+"5;:;:-%%>==-33--3332112o1X111X1X*##144*41211112222X11o12-313663>",
+"&*&*XX%%oo33-o2o2222o11oX1 o23111X1111%113112111o1o 1-%o2244:::*",
+"<4**XXXoo2332o232122o o2211X22o1o1o2##o211o12211111o11o12:54q6**",
+",#$&X o12oo33333222233o2oo11o111111%%21%11332222211X#o2-62::444:",
+"&*@& X1ooo22223312232311112111oX1Xo%1%121133121331o2XX-%22244*2:"
+};
