+++ backroom/floor3.xpm
@@ -1,0 +1,109 @@
+/* XPM */
+static char *_650504352706[] = {
+/* columns rows colors chars-per-pixel */
+"64 64 39 1 ",
+"  c #757576",
+". c #777876",
+"X c #797977",
+"o c #747678",
+"O c #777A7B",
+"+ c #7B7C7C",
+"@ c #807F7C",
+"# c #7E807E",
+"$ c #81817E",
+"% c #7D7E80",
+"& c #7D8082",
+"* c #838484",
+"= c #878887",
+"- c #898986",
+"; c #858689",
+": c #87888A",
+"> c #8B8C8C",
+", c #908F8C",
+"< c #8F908F",
+"1 c #91918E",
+"2 c #8E8E91",
+"3 c #8F9092",
+"4 c #939494",
+"5 c #999794",
+"6 c #979896",
+"7 c #989996",
+"8 c #969799",
+"9 c #979899",
+"0 c #9B9C9B",
+"q c #A09F9C",
+"w c #9FA09F",
+"e c #A0A09F",
+"r c #9E9FA0",
+"t c #A09EA2",
+"y c #9FA0A1",
+"u c #A3A3A3",
+"i c #A6A8A7",
+"p c #A8A8A6",
+"a c #AAAAAA",
+/* pixels */
+"-$5ee1,pu8>>;;::22;22>288t2832tttqu5*1851>,>5$*+$***--->>,:-:*-@",
+"01qepw0700;rr;%r0:;;:4:22284238rq0551,>071-4>$-$--,--,2,**;:>*$$",
+"55qeiay3>r3uar%r2:208;3;3tr848288,2811>0047----.@@>>-,51;*;*2;::",
+",15euuw6uu0iai%&;:243*22tit;%424,:15544ue87<1>-+@+--*>14,*%-222:",
+"--46u407quutr8rr*%*42&&:980%O>12>=,>044q70541>>-*$>,*1,*-,;2;2;*",
+"X*33661w07qu8rrr4*2282224;3*$**>>->>51-404561>>1-$;,:5---;***2>%",
+"X+=;:667t4,508rr4:;:%2;3**122:>>>,-=,-@,640803:<,-1-;2,>-*%+*>;%",
+"##:00ai0q7880943>O%o%%;3*;1:;2>;$-,--*@>15082>:--@>>>1,---%++***",
+"<:23wai950644e05;;:2;:>&t282;2;2----,,-<<161<>:=$$*-2,*>-,;*****",
+"=:;>2ia607e07ee032222;225;*8222211-*##--><<<&>:=;*$*,;%*1,;*%;**",
+"==&**yay00q00074:2:22;2r*;855524,1-$-$$<*46=>:<:;-*-;@  **%++*$+",
+"&#&%&>ay0056w9222;;:**&4;442822841,-->>>3<6<3><3,*@,,*@%%@@%+$$%",
+"<&&=&&rr00e1992::::;%%%%*;;;2;885w71$4<<0984<>>>;*$,11;$ oX%%;o ",
+"<#O#O&890w77043:;2r*%%%2%;;22222555,-5ww9u07<>>:-*@,>>>%% o%**%o",
+"#O##*;<<w0w0044>;rt:%o%%;;%%;22;<--#$www0w00>=>**%*>2,;%;%%%;;%@",
+"#XXO&440w700u0>:;:r%%%%%2;%;;2>;#.#.$wwe06w094:*:*2+;%%%22;*%2>*",
+"O##.#wwu040004=&:=>#:##&22;*%22;&..#*3400700u0>:;:q*&%%%&:&:&22:",
+"<>-#-7ww0u00-=<;&&#><<*&;+%%;;+ &.#&&;3<u0w0043>;tt>%o%&&2&&332&",
+"677,>4w68u072>>::*+1,>>&$ o%*;% <#+#O&997w06042:;20;%%%2&&&32232",
+"5771-<1<09883>>>>=#113*&ooo%%;@ <*&&&&ry00w499>>-;;;%%o%&22&3&r8",
+"55->->>>3>6>2>42>$$1<*.+%++*+%*+*##&&2ay0075q0>22;:>%;&4*441r23r",
+",1$--$$1=48:2->->----X.o**%++*%+==&&*wiyw0w0e784:2>22*20*;888322",
+",,-$##>--41>&>-:-*$$<>&*1,;*%;**=::>>aa677e78et032>32:324;;822:2",
+"$----<->1464>>:=##*=<>*<-,;$****<:<3wai777568u75:;2;::>;u381;2;2",
+"$,,$-$$<14w52>:--$>>>,,---*++***##>7wai070087084>X%+o%;2*;1;;>>:",
+"11,-,-$,45w764:<--<--2,>-*@+*2;+XO:::946q4,505004;;:;:;2=*222:2:",
+",=-,1,-154441>>1-*->>4--*>%**2>%O*<3741wt7ri00uu3;224::24:3*$;*>",
+",=<>w45www511,>*$*>>*1,--,:2;2;*=>17p670tttt88u0;@;22;*2840%+>1>",
+"<=14w<4ew45<,>-+@+>-*>11,*%:22>;115eep06ttriar*:;;>32;22rit;%324",
+"7<,11<<7w46,---X$@,>-,41-*;*2:;>150eiay3:r2tar%82;3t8;2;3tt84828",
+"ww5w<<,w51-1-$-$-*>-,,,>**;-2**$q1qep0w9rr:rr:&r0:;;:4;22248224r",
+"wwe5$1741-<<5$*#$-**--*,>,-;:**@-$5ee1>ir82:;:;:22:222289t2882tt",
+"eee5*176,-,15$=@@;**--;,>,>;;**$$$5ee,,pr82;*;%:22:22:288t2822tt",
+"0q771>>784-1>@*$-*->,>,>$*;;2;*%e5eeeeeqrr:rr:%rr:;;:8:22284248r",
+"01,81110085>,-*X@@,-;,5,*;;*2;;>555epae1:r4uar%r2;3r8;2;3tt82r28",
+",-44745eq571->-+@@;>*,,5,;$;22>;,55epe06uu0uai%:;;232*24tat:%422",
+">->>q4400074>>>-**,,;12*-,;2;2-*==46e100qupe96w0%%;32;;:880@o22>",
+">->>1,-5744842>1-@;,>1----@;*2>%+=<3061wq00u00pu8;124::24;2*$;;:",
+">>--,-$>540593:>,;,>>,,,-*@+%2;%.#*:>010e1-8080w4;;:*:;2**222;>>",
+"$-1--*$>170732--;@>>;,,-->*o+***+#>07pp000740044:O%+O%;2*:2-;2>;",
+"--,$,<-,1844>>:=@@*-,,*,-,:@****<=<3wai660668w06:&2>::2&t482;2;2",
+"11,$#$---11>*>-=-*@*,>%*2>:*@:*%*:=>3ai877w08yw08:>32:3:4*;54>>1",
+",1$$-$#1*45;>>>>>;@,*%@ $*%%@$%%;=&;&yayw000y788:2222&2r*;888424",
+"41,,->>,438>>>33-*@,>* @%%o;%@*%&##&&>aw0w46y9>>3&;2&;%2;4418248",
+"7771-<<<09882>2>>:@,22**  o&%;oo<&#&&&www0e170>>2::&%%%%&;;;3;88",
+"775,-4ww8u0822:;:@@2-,>%% o%**$o<&OO#&864w05043=;3r;%%%:&;;:223:",
+"<>=#-7ww7u00>;,*$&%>,2*%;%+%;;% &###&:<<ewe7w44>;yy>%o%*;:*%:22;",
+".##X#wweq80q81;*>;2@;*%%22;**22;$X.#&38ww55we7<=;>r&%%%%2:%;;22:",
+"#.X#&88ww47we7<-:>0&:&&%&&&2*22:+X#X#0wu78qt81;*>;>&-##&>2:*%22*",
+"#X##&&3<wwwww43>&yu>%o%&&3&&;22*<--#=0u07u052;>;&&&:<>=&:O%%::@ ",
+"<#X#O&8w8w67y3<=;2r:%&%3&&&3222>675<-4000u981>>-=##<>>>%& o%**$o",
+"<&&&&#www6w199>>>>&&%&oo&:::22847771->4190881>>>>-#143;$ooO%%*Oo",
+"&#&&&3ay0w5500>22&:>%;&2:833934841,$->->42822>22>*$1>=.%%%%*+%@&",
+">-&&&yayw7w0w088:2232;3r&:889424,1,$-*$4*48:>>>>-**--X. $%%@+*%%",
+"=>&>3ia657e06yw08:232;228&;93>>311,$@@-->1>>*>-:-=$*1-&*2>:;@:**",
+">=<4wai677486006:&>>>;2;y883:2:2,--->,->14742-::##*-,>*,-2:$$*;$",
+"+&>90pa7w0740044>o@OX+%3*:2;:>,;$-<--##<25t82::--@>>>2,---$%%;$*",
+".+*;-040e1,505004::>*>;2**221;>:,,-=,-#,54q493:,,;,>;,,,-*@@%:;@",
+"+*449810w00e00pu3:224:224*3;**:>,-<-5,-1555432,,-$;,>1---,@**2>%",
+"=:17u400weep80r0*+:42*&2850%O:3>-$<<w45w5t54,2>*$$>,;1,;-,:2;2;&",
+"<15wuur9utriau%>;:242*22tpt;+424<-14544et15,,,*@@@>>*,,1,*@;22>;",
+"66wepat32r2ra0%r2;208:2;4tt548285<<7,<,wt58,,**@$@>>-,14;;;*2;;>",
+"q1eeuqrrrr;rr:%rr;;;24;222843487www5<>>w51;5-@*@-*>--,,>**;>2***",
+"$$50u>:ur82:;:;&22;2;2240t2833weewe5$<54,>,,5@;@@;**--;,>,-;;**%"
+};
