+++ backroom/glass.xpm
@@ -1,0 +1,104 @@
+/* XPM */
+static char *_650504356964[] = {
+/* columns rows colors chars-per-pixel */
+"64 64 34 1 ",
+"  c #373737",
+". c gray22",
+"X c #393939",
+"o c #3A3A3A",
+"O c gray23",
+"+ c #3C3C3C",
+"@ c gray24",
+"# c #3E3E3E",
+"$ c #3F3F3F",
+"% c gray25",
+"& c #414141",
+"* c gray26",
+"= c #434343",
+"- c #444444",
+"; c gray27",
+": c #464646",
+"> c gray28",
+", c #484848",
+"< c #494949",
+"1 c gray29",
+"2 c #4B4B4B",
+"3 c #4C4C4C",
+"4 c gray30",
+"5 c #4E4E4E",
+"6 c gray31",
+"7 c #505050",
+"8 c #515151",
+"9 c gray32",
+"0 c #535353",
+"q c gray33",
+"w c #555555",
+"e c #565656",
+"r c gray34",
+"t c None",
+/* pixels */
+".t.t.t.t.t t t t t t t. . .....t.t.t.t.t t t t t t t.t.t.t.t.t.t",
+"tttttttttttttttttttttttttttttttttttttt.t.t.t.t t t t t t t t t t",
+".t t.t.t.t t.t.t.t t.t.t.t t.t.t.t.t.t.t t t t t t t.t.t.t.t.t.t",
+"tttttttttttttttttttttttttttttttttttttt.t.t.t.t t t t t t t t t t",
+".t t.t.t.t t.t.t.t.t.t.t.t.t.t.t.t.t.t@t.t.t.t.t t t.t.t.t.t.t.t",
+"tttttttttttttttttttttttttttttttttttttt.t.t@t.t.t t t t t t t t t",
+".t t.t.t.t t.t.t.t.t.t@t@t@t@t@t%t@t@t@t@t.t.t.t t.t.t.t.t.t.t.t",
+"tttttttttttttttttttttttttttttttttttttt@t@t@t@t.t.t t t t t t t t",
+".t t.t.t.t t.t.t.t@t@t@t%t%t%t%t%t%t%t%t%t%t@t.t.t.t.t.t.t.t.t.t",
+"tttttttttttttttttttttttttttttttttttttt%t%t%t%t@t@t.t t.t t t t t",
+".t t.t.t.t t.t.t@t@t%t%t-t-t-t-t1t-t-t-t-t%t%t@t@t.t.t.t.t.t.t.t",
+"tttttttttttttttttttttttttttttttttttttt1t1t-t%t%t@t@t.t.t t t t t",
+".t t.t.t.t.t.t@t%t%t-t-t1t1t1t3t3t3t1t1t-t-t-t%t@t@t@t.t.t.t.t.t",
+"tttttttttttttttttttttttttttttttttttttt3t1t1t-t%t@t@t.t.t t t t t",
+".t t.t.t.t.t@t%t%t-t1t1t3t3t3t7t7t3t3t3t3t3t1t-t%t@t@t.t.t.t.t.t",
+"tttttttttttttttttttttttttttttttttttttt7t3t3t1t1t-t%t@t@t.t t t t",
+".t t.t.t.t@t%t-t-t1t1t3t7t7t7t7t7t7t7t7t7t3t1t1t-t%t@t@t.t.t.t.t",
+"ttttttttttttttttttttttttttttttttttttttwt7t7t3t1t1t-t%t@t.t.t.t.t",
+".t t.t.t.t@t%t-t1t3t3t7t7twt7twtwtwtwt7t7t7t3t3t1t-t%t@t@t.t.t.t",
+"ttttttttttttttttttttttttttttttttttttttwt7t7t7t3t1t-t%t%t@t.t t t",
+".t t.t.t@t%t-t-t3t3t7t7t7twtwtwtwtwtwtwtwt7t7t3t1t-t-t%t@t.t.t.t",
+"ttttttttttttttttttttttttttttttttttttttwtwtwtwt7t1t1t-t%t@t.t.t t",
+".t t.t@t%t%t-t1t3t7t7twtwtwtwtwtwtwtwtwtwt7t7t7t3t1t-t%t@t@t.t.t",
+"ttttttttttttttttttttttttttttttttttttttwtwtwtwt7t3t3t1t-t%t.t.t t",
+".t t.t@t%t-t1t3t7t7t7twtwtwtwtrtwtrtwtrtwtwtwt7t3t3t1t-t@t@t.t.t",
+"ttttttttttttttttttttttttttttttttttttttwtrtwtwt7t7t1t1t%t%t.t.t.t",
+".t t.t@t%t-t1t3t7twtwtwtwtrtrtrtrtrtrtrtwtwtwt7t7t3t1t-t%t@t.t.t",
+"ttttttttttttttttttttttttttttttttttttttwtrtwtwtwt7t3t1t-t%t@t.t.t",
+".t.t.t@t%t-t3t3t7twtwtwtwtrtrtrtrtrtrtrtrtwtwt7t7t3t1t-t%t@t.t.t",
+"ttttttttttttttttttttttttttttttttttttttrtwtrtwtwt7t7t3t-t%t@t.t.t",
+".t.t.t@t%t-t3t7t7twtwtwtrtrtrtrtrtrtrtrtrtwtwtwt7t3t3t1t%t@t.t.t",
+"ttttttttttttttttttttttttttttttttttttttwtwtrtrtwt7t7t3t-t%t@t.t.t",
+".t.t.t@t%t1t3t7t7twtwtwtrtrtrtrtrtrtrtrtrtwtwtwt7t7t3t-t%t@t.t.t",
+"ttttttttttttttttttttttttttttttttttttttrtrtrtwt7t7t7t1t1t%t@t.t.t",
+".t.t.t@t%t-t1t3t7twtwtwtrtrtrtrtrtrtrtrtwtrtwtwt7t3t1t-t%t@t.t.t",
+"ttttttttttttttttttttttttttttttttttttttwtrtwtrtwt7t3t1t1t%t@t.t.t",
+".t.t.t%t%t-t1t3t7t7twtwtwtrtrtrtrtrtrtrtwtrtwtwt7t3t1t-t%t@t@t.t",
+"ttttttttttttttttttttttttttttttttttttttrtwtwtwt7t7t3t1t-t%t@t.t.t",
+".t.t.t@t%t-t1t3t7twtwtwtwtrtrtrtrtrtrtwtwtwtwt7t7t3t1t-t%t.t.t.t",
+"ttttttttttttttttttttttttttttttttttttttrtwtwtwt7t7t3t-t-t%t@t.t t",
+".t.t.t@t%t-t1t1t7t7twtwtwtwtwtwtrtrtwtrtwtwtwt7t3t1t1t%t@t.t.t.t",
+"ttttttttttttttttttttttttttttttttttttttwtwtwt7t7t3t1t-t%t@t@t.t.t",
+".t.t.t.t@t%t-t1t3t7t7t7twtwtwtwtwtwtwtwtwt7twt7t3t1t-t%t@t@t.t.t",
+"ttttttttttttttttttttttttttttttttttttttwtwt7t7t3t3t1t-t%t@t.t t.t",
+".t.t.t.t@t%t-t1t3t3t7t7twtwtwtwtwtwtwtwtwt7t7t3t1t-t%t%t@t.t.t.t",
+"tttttttttttttttttttttttttttttttttttttt7t7t7t7t3t1t-t%t@t@t.t t t",
+".t.t.t.t.t@t%t-t1t3t3t7t7twt7twtwt7twt7t7t7t3t3t1t-t%t@t.t.t.t.t",
+"tttttttttttttttttttttttttttttttttttttt7t7t7t3t1t-t-t%t@t.t.t t t",
+".t.t.t.t.t@t@t%t-t1t3t3t3t7t7t7t7t7t7t7t3t3t1t1t-t%t@t@t.t.t.t.t",
+"tttttttttttttttttttttttttttttttttttttt3t3t1t1t-t%t%t@t.t.t.t t t",
+".t.t.t.t.t.t.t%t%t-t1t1t3t3t3t3t7t3t7t3t3t1t-t-t%t%t@t.t.t.t.t.t",
+"tttttttttttttttttttttttttttttttttttttt1t1t1t-t%t%t@t.t.t.t.t t t",
+".t t.t.t.t.t.t@t@t%t-t-t1t1t1t3t3t3t3t1t1t-t-t%t@t@t.t.t.t.t.t.t",
+"tttttttttttttttttttttttttttttttttttttt-t%t-t%t@t@t.t.t.t t t t t",
+".t t.t.t.t.t.t.t.t@t%t%t%t-t-t-t-t-t-t-t%t%t%t@t.t.t.t.t.t.t.t.t",
+"tttttttttttttttttttttttttttttttttttttt-t%t%t@t@t.t@t.t.t t t t t",
+".t t.t.t.t.t.t.t.t.t@t@t%t%t%t%t%t%t%t%t@t@t.t.t.t t.t.t.t.t.t.t",
+"tttttttttttttttttttttttttttttttttttttt@t@t@t@t.t.t t t t t t t t",
+".t t.t.t.t.t.t.t.t.t.t@t@t@t@t@t%t@t@t@t@t.t.t.t t t.t.t.t.t.t.t",
+"tttttttttttttttttttttttttttttttttttttt.t.t.t.t.t t t t t t t t t",
+".t t.t.t.t t.t.t.t t.t.t.t.t.t.t.t.t.t@t.t.t t t t t.t.t.t.t.t.t",
+"tttttttttttttttttttttttttttttttttttttt.t.t.t.t t t t t t t t t t",
+".t t.t.t.t.t.t.t.t t.t.t.t.t.t.t.t.t.t t t t t t t t.t.t.t.t.t.t",
+"tttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttt"
+};
