+++ backroom/light1.xpm
@@ -1,0 +1,108 @@
+/* XPM */
+static char *_650504364512[] = {
+/* columns rows colors chars-per-pixel */
+"64 64 38 1 ",
+"  c #F5F98E",
+". c #EEF692",
+"X c #EFF793",
+"o c #F0F894",
+"O c #F4FA9D",
+"+ c #F2F9A2",
+"@ c #F4FBA4",
+"# c #F5FCA5",
+"$ c #F6FDA6",
+"% c #F7FBAB",
+"& c #F2F9B0",
+"* c #F3FAB1",
+"= c #F4FBB2",
+"- c #F5FCB3",
+"; c #F6FDB4",
+": c #F2F9BD",
+"> c #F3FABE",
+", c #F5FCC0",
+"< c #F6FDC1",
+"1 c #F4FACC",
+"2 c #F5FBCD",
+"3 c #F6FCCE",
+"4 c #F7FDCF",
+"5 c #FDFFC9",
+"6 c #F8FED0",
+"7 c #F5FAD8",
+"8 c #F6FBD9",
+"9 c #F8FDDB",
+"0 c #F9FEDC",
+"q c #F6FAE7",
+"w c #F7FBE8",
+"e c #F8FCE9",
+"r c #F9FDEA",
+"t c #FAFEEB",
+"y c #FBFDF7",
+"u c #FCFEF8",
+"i c #FDFFF9",
+"p c #FDFFFC",
+/* pixels */
+"   .oO+@$%&;;>><,<5141616466668777777746661611615,,,:--=&%$@@ooo",
+"   ..++@@%&&&>>>,,5141416666647777777766616111115,,,:---=%@@+oXX",
+"  ...+$$$&&;;>><<511116664877788787878874666116115,,>:--=&%@@Ooo",
+"....@@@%&&;>>><551411466877778088080808778666661115,,,,>--&%@@++",
+".oo+$$%&&;><<5111166487880999900009909988088886461115<<,,--&%@@@",
+"OO++$%=&>>>511364488899900qwqqwwwwqwqq09089888878661115<,>>-&%$@",
+"+++$%=&;><<1366348788000qqwwqweweewwwwwqqq00088788466115,,>-=&%%",
+"@@$%&&;<<5133388899900qqeweeeeetteweeewewwwwww00088466111,,,--&-",
+"$$$&&>><>13367898900qqwwteeeetteeteteteeeewwwww00088866115<,>---",
+"%%&&;,,51363888990qqweteettetetetettteteteeewtwww0088884611<,>::",
+"&&&;>,<1366788900qwwetetteyyutuuuuuuyytettettewtww0808886115,,>>",
+";&&>,5136388990wwweweetettyuuuuuuuuuuuuuuutttteewwww008886615,,,",
+";&;><116378890wqewetteeeuuuuuuupuupuuuuuyyttuuttwtwww098866115<,",
+">>>,513687999wwwettettuuuuuupuuuuuuuupupuuuuyyteteewww0988611155",
+">>,,13388990wwweteeeuuuuuuuuuuppupppuupuuuuuuuyytttewww088861411",
+",,,53368899qwweeetuyuuuuupuupuuuppupppuuuuuuuuyyttetwwww99866114",
+",,<13669000qwwetuuuuuupuupuupppppppuuppppppuupuuuytttetw00884111",
+",,51368880wwwwetyuuuuuuuuupppupppuppppppuuuuuuuuyytteteeq0986464",
+"551138890weeeetuyuuuuuppppppppppppppppppppppuppuuuyytttewq088664",
+"111368890wweeetuuuuupuppppppppppppppppppppupppuuuuyyetewwq097866",
+"11136799wwweetyyuuuuupupppppppppppppppppppppppuuuuuuueetewq97846",
+"31368990wweeteyuuupuuuupuuupppppppppppppppuuupupuuuuuttewwq08878",
+"3336890wweettuuuuuppuppppppppppppppppppppppuppppppupuueteewq0878",
+"3363799wwetteuuupuppppppppppppppppppppppppppppppppuuuueteeww9988",
+"663899wwwettyyyuupupppppuupppppppppppppppppppuuupuupuuyyteee0988",
+"633889wwwettuuuuuuuppuupppppppppppppppppppppppupuuuuuuyyttww0098",
+"66880wweetuuuuuuupppppppppppppppppppppppppppppuuppuuuuuuutetw009",
+"63880wweteyyuupupppppppppppppppppppppppppppppppppppuuuyyueteq099",
+"66789wweetuuyuupppppppppppppppppppppppppppppppppuuupupuuutteq099",
+"66788weeeteuuuuuppppppppppppppppppppppppppppppppppuuuuuuuewwq099",
+"87880wwettuuuuppuuppppppppppppppppppppppppppppppppppupuyyeeewq00",
+"87890weettuuuupppppppppppppppppppppppppppppppppppppuuuuuuteewq00",
+"77890eetttuuuppuppppppppppppppppppppppppppppppppppppuuuuuetewq99",
+"77700wweteuuuupppppppppppppppppppppppppppppppppppppppuuuutetww09",
+"77789qweetuuupuppupppppppppppppppppppppppppppppppppuuuuyuteeqw00",
+"77899qwettuuuuppppppppppppppppppppppppppppppppppppppuuuuueetee00",
+"77789qeeetuuuuupupppppppppppppppppppppppppppppppppppuuuuuetewq99",
+"77809wwettuupupppppppppppppppppppppppppppppppppppppppuuuuteeqq00",
+"667788weeteuuuupupppppppppppppppppppppppppppppppuupuuuuteeewe099",
+"648899wwteeuuuuuppppppppppppppppppppppppppppppppppuuuuueeteqw099",
+"446889qeweeuyuuupuppppppppppppppppppppppppppppuuppuuuuuttteww999",
+"444809qwtetuyuuupppppppppppppppppppppppppppppppppuuuuuutteewq999",
+"1646999qweeteuuuuuupppupppppppppppppppppppppppuuuuuuuuteteww9099",
+"6146899qtetttuuupupppuupupppppppppppppppppppuppupuuuyyttete09889",
+"4166709qweeeyyuuuuppuupppppppppppppppppppppuppupupuuuueeeee99987",
+"1116789qweetyyuupupppppppppppppppppppppppppuppuuuuuuuutewqw99977",
+"11114788qeetetyyuuuuuuppuupuppppppppppppppuuupuuuuuuttewwq997846",
+"11114890wwtettuuuuupuppuupppppppppppppppupupuuuuuuuueteeeq087644",
+"551168880wweteteyyuuuupupupppppppppppppuppupuuuuyyeetewww9997616",
+"<<51168808wwwettuuuuuupppuupuppppppppppupuuuuuuuyyteetewe9974611",
+",<<51668008wwetettuuuupuuuuupuuppuupupupuuuuuuyyttteeeqq09986641",
+">,<<16668800wtetteyuuuuuuuupuuppppppppuupuuuuuuuttetweee99844111",
+">::,51168808weetetetuuuuuuuupuupuuupupuuuuuuyyetteewweq997861155",
+";-:,<11668008wwwtetteeuuuuuuuuuupuuuuuuuuuuuuutteteeww99986115<<",
+";=;:,516188800qwtwetetttuuuuuuuuuuuuuuuuuuwetttetewqq99974615,,,",
+"==;>><511688880wwwweteetuuuuuuuuuuuuuutetteteeweewww99986661,,,,",
+"&==;>,<116668800wqwweweettuuutuuyyuuyuttttetewqeqq090808611,,>>>",
+"%%&=;>><5116688890qqwwweteeeteetteteeteeeeteewqq00998846115<>>;=",
+"@$%=->>,<11666888900wwwweteteteeeettetteewwewq999988716115<>>==;",
+"@@@%=-->,<511114889099qqwwtwewweetwetwwweqqq099989846611,,,>====",
+"++@@%==-:,,511666688889009wqwqwwwwwwwwqqqq0008888644115<>>>==&%%",
+"ooO@@%==->,<511116688898090900qqqqqwww9999998886161115<<>>;==@@@",
+"XXo+@$%&=>>,<5164166688808999900900000999988886466115<,<>;==%@@@",
+"Xoo+@$%==:><<5111116478888999900000000999908884461415<,>>;==%%@@"
+};
