+++ backroom/wall0.xpm
@@ -1,0 +1,72 @@
+/* XPM */
+static char *_650504297247[] = {
+/* columns rows colors chars-per-pixel */
+"64 64 2 1 ",
+"  c #DCE5BC",
+". c #F5FFD1",
+/* pixels */
+"    ....    ............    ....    ....    ............    ....",
+"    ....    ............    ....    ....    ............    ....",
+"    ....    ............    ....    ....    ............    ....",
+"    ....    ............    ....    ....    ............    ....",
+"....        ............        ....        ............        ",
+"....        ............        ....        ............        ",
+"....        ............        ....        ............        ",
+"....        ............        ....        ............        ",
+"........    ....    ....    ............    ....    ....    ....",
+"........    ....    ....    ............    ....    ....    ....",
+"........    ....    ....    ............    ....    ....    ....",
+"........    ....    ....    ............    ....    ....    ....",
+"........        ....        ............        ....        ....",
+"........        ....        ............        ....        ....",
+"........        ....        ............        ....        ....",
+"........        ....        ............        ....        ....",
+"    ....    ............    ....    ....    ............    ....",
+"    ....    ............    ....    ....    ............    ....",
+"    ....    ............    ....    ....    ............    ....",
+"    ....    ............    ....    ....    ............    ....",
+"....        ............        ....        ............        ",
+"....        ............        ....        ............        ",
+"....        ............        ....        ............        ",
+"....        ............        ....        ............        ",
+"........    ....    ....    ............    ....    ....    ....",
+"........    ....    ....    ............    ....    ....    ....",
+"........    ....    ....    ............    ....    ....    ....",
+"........    ....    ....    ............    ....    ....    ....",
+"........        ....        ............        ....        ....",
+"........        ....        ............        ....        ....",
+"........        ....        ............        ....        ....",
+"........        ....        ............        ....        ....",
+"    ....    ............    ....    ....    ............    ....",
+"    ....    ............    ....    ....    ............    ....",
+"    ....    ............    ....    ....    ............    ....",
+"    ....    ............    ....    ....    ............    ....",
+"....        ............        ....        ............        ",
+"....        ............        ....        ............        ",
+"....        ............        ....        ............        ",
+"....        ............        ....        ............        ",
+"........    ....    ....    ............    ....    ....    ....",
+"........    ....    ....    ............    ....    ....    ....",
+"........    ....    ....    ............    ....    ....    ....",
+"........    ....    ....    ............    ....    ....    ....",
+"........        ....        ............        ....        ....",
+"........        ....        ............        ....        ....",
+"........        ....        ............        ....        ....",
+"........        ....        ............        ....        ....",
+"    ....    ............    ....    ....    ............    ....",
+"    ....    ............    ....    ....    ............    ....",
+"    ....    ............    ....    ....    ............    ....",
+"    ....    ............    ....    ....    ............    ....",
+"....        ............        ....        ............        ",
+"....        ............        ....        ............        ",
+"....        ............        ....        ............        ",
+"....        ............        ....        ............        ",
+"........    ....    ....    ............    ....    ....    ....",
+"........    ....    ....    ............    ....    ....    ....",
+"........    ....    ....    ............    ....    ....    ....",
+"........    ....    ....    ............    ....    ....    ....",
+"........        ....        ............        ....        ....",
+"........        ....        ............        ....        ....",
+"........        ....        ............        ....        ....",
+"........        ....        ............        ....        ...."
+};
