+++ backroom/wall1.xpm
@@ -1,0 +1,73 @@
+/* XPM */
+static char *_650504301900[] = {
+/* columns rows colors chars-per-pixel */
+"64 64 3 1 ",
+"  c #EBF4C8",
+". c #EFF9CC",
+"X c #F5FFD1",
+/* pixels */
+"    XXXX    XXXXXXXXXXXX....XXXX    XXXX....XXXXXXXXXXXX    XXXX",
+"    XXXX    XXXXXXXXXXXX....XXXX    XXXX....XXXXXXXXXXXX    XXXX",
+"    XXXX    XXXXXXXXXXXX....XXXX    XXXX....XXXXXXXXXXXX    XXXX",
+"    XXXX    XXXXXXXXXXXX....XXXX    XXXX....XXXXXXXXXXXX    XXXX",
+"XXXX        XXXXXXXXXXXX....    XXXX    ....XXXXXXXXXXXX        ",
+"XXXX        XXXXXXXXXXXX....    XXXX    ....XXXXXXXXXXXX        ",
+"XXXX        XXXXXXXXXXXX....    XXXX    ....XXXXXXXXXXXX        ",
+"XXXX        XXXXXXXXXXXX....    XXXX    ....XXXXXXXXXXXX        ",
+"XXXXXXXX    XXXX    XXXX....XXXXXXXXXXXX....XXXX    XXXX    XXXX",
+"XXXXXXXX    XXXX    XXXX....XXXXXXXXXXXX....XXXX    XXXX    XXXX",
+"XXXXXXXX    XXXX    XXXX....XXXXXXXXXXXX....XXXX    XXXX    XXXX",
+"XXXXXXXX    XXXX    XXXX....XXXXXXXXXXXX....XXXX    XXXX    XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX",
+"    XXXX    XXXXXXXXXXXX....XXXX    XXXX....XXXXXXXXXXXX    XXXX",
+"    XXXX    XXXXXXXXXXXX....XXXX    XXXX....XXXXXXXXXXXX    XXXX",
+"    XXXX    XXXXXXXXXXXX....XXXX    XXXX....XXXXXXXXXXXX    XXXX",
+"    XXXX    XXXXXXXXXXXX....XXXX    XXXX....XXXXXXXXXXXX    XXXX",
+"XXXX        XXXXXXXXXXXX....    XXXX    ....XXXXXXXXXXXX        ",
+"XXXX        XXXXXXXXXXXX....    XXXX    ....XXXXXXXXXXXX        ",
+"XXXX        XXXXXXXXXXXX....    XXXX    ....XXXXXXXXXXXX        ",
+"XXXX        XXXXXXXXXXXX....    XXXX    ....XXXXXXXXXXXX        ",
+"XXXXXXXX    XXXX    XXXX....XXXXXXXXXXXX....XXXX    XXXX    XXXX",
+"XXXXXXXX    XXXX    XXXX....XXXXXXXXXXXX....XXXX    XXXX    XXXX",
+"XXXXXXXX    XXXX    XXXX....XXXXXXXXXXXX....XXXX    XXXX    XXXX",
+"XXXXXXXX    XXXX    XXXX....XXXXXXXXXXXX....XXXX    XXXX    XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX",
+"    XXXX    XXXXXXXXXXXX....XXXX    XXXX....XXXXXXXXXXXX    XXXX",
+"    XXXX    XXXXXXXXXXXX....XXXX    XXXX....XXXXXXXXXXXX    XXXX",
+"    XXXX    XXXXXXXXXXXX....XXXX    XXXX....XXXXXXXXXXXX    XXXX",
+"    XXXX    XXXXXXXXXXXX....XXXX    XXXX....XXXXXXXXXXXX    XXXX",
+"XXXX        XXXXXXXXXXXX....    XXXX    ....XXXXXXXXXXXX        ",
+"XXXX        XXXXXXXXXXXX....    XXXX    ....XXXXXXXXXXXX        ",
+"XXXX        XXXXXXXXXXXX....    XXXX    ....XXXXXXXXXXXX        ",
+"XXXX        XXXXXXXXXXXX....    XXXX    ....XXXXXXXXXXXX        ",
+"XXXXXXXX    XXXX    XXXX....XXXXXXXXXXXX....XXXX    XXXX    XXXX",
+"XXXXXXXX    XXXX    XXXX....XXXXXXXXXXXX....XXXX    XXXX    XXXX",
+"XXXXXXXX    XXXX    XXXX....XXXXXXXXXXXX....XXXX    XXXX    XXXX",
+"XXXXXXXX    XXXX    XXXX....XXXXXXXXXXXX....XXXX    XXXX    XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX",
+"    XXXX    XXXXXXXXXXXX....XXXX    XXXX....XXXXXXXXXXXX    XXXX",
+"    XXXX    XXXXXXXXXXXX....XXXX    XXXX....XXXXXXXXXXXX    XXXX",
+"    XXXX    XXXXXXXXXXXX....XXXX    XXXX....XXXXXXXXXXXX    XXXX",
+"    XXXX    XXXXXXXXXXXX....XXXX    XXXX....XXXXXXXXXXXX    XXXX",
+"XXXX        XXXXXXXXXXXX....    XXXX    ....XXXXXXXXXXXX        ",
+"XXXX        XXXXXXXXXXXX....    XXXX    ....XXXXXXXXXXXX        ",
+"XXXX        XXXXXXXXXXXX....    XXXX    ....XXXXXXXXXXXX        ",
+"XXXX        XXXXXXXXXXXX....    XXXX    ....XXXXXXXXXXXX        ",
+"XXXXXXXX    XXXX    XXXX....XXXXXXXXXXXX....XXXX    XXXX    XXXX",
+"XXXXXXXX    XXXX    XXXX....XXXXXXXXXXXX....XXXX    XXXX    XXXX",
+"XXXXXXXX    XXXX    XXXX....XXXXXXXXXXXX....XXXX    XXXX    XXXX",
+"XXXXXXXX    XXXX    XXXX....XXXXXXXXXXXX....XXXX    XXXX    XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX"
+};
