+++ backroom/wall3.xpm
@@ -1,0 +1,85 @@
+/* XPM */
+static char *_650504308639[] = {
+/* columns rows colors chars-per-pixel */
+"64 64 15 1 ",
+"  c #5F6351",
+". c #616553",
+"X c #636855",
+"o c #81856D",
+"O c #838870",
+"+ c #868C72",
+"@ c #AEB494",
+"# c #B5BD9A",
+"$ c #BFC7A3",
+"% c #C2CBA6",
+"& c #C7D0AA",
+"* c #DCE5BC",
+"= c LightGray",
+"- c #F5FFD1",
+"; c white",
+/* pixels */
+"****----****------------****----****----****------------****----",
+"****----****------------****----****----****------------****----",
+"****----****------------****----****----****------------****----",
+"****----****------------****----****----****------------****----",
+"----********------------********----********------------********",
+"----********------------********----********------------********",
+"----********------------********----********------------********",
+"----********------------********----********------------********",
+"--------****----****----****------------****----****----****----",
+"--------****----****----****------------****----****----****----",
+"--------****----****----****------------****----****----****----",
+"--------****----****----****------------****----****----****----",
+"--------********----********------------********----********----",
+"--------********----********------------********----********----",
+"--------********----********------------********----********----",
+"--------********----********------------********----********----",
+"****----****------------****----****----****------------****----",
+"****----****------------****----****----****------------****----",
+"****----****------------****----****----****------------****----",
+"****----****------------****----****----****------------****----",
+"----****================********----********------------********",
+"----****================********----********------------********",
+"----****================********----********------------********",
+"----****================********----********------------********",
+"--------====;;;;;;;;====****------------****----****----****----",
+"--------====;;;;;;;;====****------------****----****----****----",
+"--------====;;;;;;;;====****------------****----****----****----",
+"--------====;;;;;;;;====****------------****----****----****----",
+"--------====;;;;;;;;====****------------********----********----",
+"--------====;;;;;;;;====****------------********----********----",
+"--------====;;;;;;;;====****------------********----********----",
+"--------====;;;;;;;;====****------------********----********----",
+"****----====;;;;;;;;====****----****----****------------****----",
+"****----====;;;;;;;;====****----****----****------------****----",
+"****----====;;;;;;;;====****----****----****------------****----",
+"****----====;;;;;;;;====****----****----****------------****----",
+"----****================********----********------------********",
+"----****================********----********------------********",
+"----****================********----********------------********",
+"----****================********----********------------********",
+"--------****----****----****------------****----****----****----",
+"--------****----****----****------------****----****----****----",
+"--------****----****----****------------****----****----****----",
+"--------****----****----****------------****----****----****----",
+"--------********----********------------********----********----",
+"--------********----********------------********----********----",
+"--------********----********------------********----********----",
+"--------********----********------------********----********----",
+"$$$$&&&&$$$$&&&&--------****----****----****----&&&&&&&&****&&&&",
+"$$$$&&&&$$$$&&&&--------****----****----****----&&&&&&&&****&&&&",
+"$$$$&&&&$$$$&&&&--------****----****----****----&&&&&&&&****&&&&",
+"$$$$&&&&$$$$&&&&--------****----****----****----&&&&&&&&****&&&&",
+"&&&&@@@@@@@@############%%%%$$$$----********############@@@@$$$$",
+"&&&&@@@@@@@@############%%%%$$$$----********############@@@@$$$$",
+"&&&&@@@@@@@@############%%%%$$$$----********############@@@@$$$$",
+"&&&&@@@@@@@@############%%%%$$$$----********############@@@@$$$$",
+"++++++++oooo++++oooo++++OOOO++++++++++++OOOO++++oooo++++oooo++++",
+"++++++++oooo++++oooo++++OOOO++++++++++++OOOO++++oooo++++oooo++++",
+"++++++++oooo++++oooo++++OOOO++++++++++++OOOO++++oooo++++oooo++++",
+"++++++++oooo++++oooo++++OOOO++++++++++++OOOO++++oooo++++oooo++++",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX",
+"XXXXXXXX        XXXX    ....XXXXXXXXXXXX....    XXXX        XXXX"
+};
