+++ include/defs__texture.h
@@ -1,14 +1,19 @@
 #ifndef DEFS__TEXTURE_H
 #define DEFS__TEXTURE_H
 
-#define TEX_AMOUNT 8
-#define NORTHWALL 0
-#define SOUTHWALL 1
-#define EASTWALL 2
-#define WESTWALL 3
+#define TEX_AMOUNT 4
+#define TEX_WIDTH 64
+#define TEX_HEIGHT 64
+#define WNORTH 0
+#define WSOUTH 1
+#define WEAST 2
+#define WWEST 3
 #define FLOOR 4
 #define CEILING 5
-#define FLOOR2 6
-#define CEILING2 7
+#define CEILING2 6
+
+#define FLOOR_PATH "backroom/floor1.xpm"
+#define CEILING_PATH "backroom/ceiling0.xpm"
+#define CEILING2_PATH "backroom/light0.xpm"
 
 #endif /* DEFS__TEXTURE_H */
